+++ OOD/src/main/java/com/distrimind/ood/OOD.java
@@ -65,6 +65,17 @@ public class OOD {
 			VERSION.addCreator(new Person("mahdjoub", "jason"))
 					.addDeveloper(new PersonDeveloper("mahdjoub", "jason", "2013-04-01"))
 					.addDescription(
+							new Description((short)3, (short)1, (short)17, Version.Type.STABLE, (short)0, "2022-02-08")
+									.addItem(INTERNAL_CHANGE, "Update Utils to 5.22.2 STABLE")
+									.addItem(INTERNAL_CHANGE, "Update HSQLDB to 2.6.1")
+									.addItem(INTERNAL_CHANGE, "Update H2 to 2.1.210")
+									.addItem(INTERNAL_CHANGE, "Update mysql-connector-java 8.0.28")
+									.addItem(INTERNAL_CHANGE, "Update postgresql 42.3.2")
+									.addItem(INTERNAL_CHANGE, "Minimum Android version to use H2 is Android 8.0 or API26")
+									.addItem(BUG_FIX, "Check if certificate is given when initialising database configurations that are synchronized with central database backup")
+									.addItem(INTERNAL_CHANGE, "Update URLs")
+					)
+					.addDescription(
 							new Description((short)3, (short)1, (short)16, Version.Type.STABLE, (short)0, "2022-01-24")
 									.addItem(INTERNAL_CHANGE, "Update Utils to 5.22.1 STABLE")
 					)